requirements.txt
fpdf2==2.7.3
lmfit==1.1.0
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
PyYAML==6.0
schema==0.7.5
scipy==1.10.1
seaborn==0.12.2
xlrd==2.0.1
